:recycle: Extract dummy_publisher_test.go
